+++ Cargo.toml
@@ -1,29 +1,32 @@
-[package]
-name = "alloy-suave"
-version = "0.1.0"
-edition = "2021"
+[workspace]
+members = [
+    "crates/*",
+    "examples/",
+]
+resolver = "2"
 
-[dependencies]
+[workspace.package]
+version = "0.0.0"
+edition = "2021"
+rust-version = "1.76"
+license = "GPL-3.0"
+exclude = [ "examples/" ]
+
+[workspace.dependencies]
+suave-alloy-network = { path = "crates/network" }
+suave-alloy-types = { path = "crates/types" }
+suave-alloy = { path = "crates/suave-alloy" }
+
 alloy = { git = "https://github.com/alloy-rs/alloy", rev = "bfd0fda", features = [
     "contract",
     "network",
-    "node-bindings",
     "providers",
     "provider-http",
-    "provider-ipc",
-    "provider-ws",
     "rpc-client",
-    "rpc-client-ipc",
-    "rpc-client-ws",
     "rpc-types-eth",
     "rpc-types-trace",
     "signers",
-    "signer-keystore",
-    "signer-ledger",
-    "signer-mnemonic",
-    "signer-trezor",
     "signer-wallet",
-    "signer-yubihsm",
     "consensus",
     "eips",
     "serde",
@@ -31,8 +34,6 @@ alloy = { git = "https://github.com/alloy-rs/alloy", rev = "bfd0fda", features =
 alloy-rlp = { version = "0.3.4", features = ["derive"] }
 async-trait = "0.1.79"
 eyre = "0.6.12"
-reqwest = "0.12.3"
 serde = { version = "1.0.197", features = ["serde_derive", "derive"] }
 serde_json = "1.0.115"
 tokio = { version = "1.37.0", features = ["full"] }
-url = "2.5.0"
