+++ crates/network/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "suave-alloy-network"
+
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+
+[dependencies]
+suave-alloy-types = { workspace = true, default-features = false }
+
+alloy = { git = "https://github.com/alloy-rs/alloy", rev = "bfd0fda", features = [
+    "network",
+    "providers",
+    "provider-http",
+    "rpc-types-trace",
+    "signers",
+] }
+
+reqwest = "0.12.3"
+async-trait.workspace = true
+
+[dev-dependencies]
+tokio.workspace = true
+eyre.workspace = true
+alloy = { git = "https://github.com/alloy-rs/alloy", rev = "bfd0fda", features = [
+    "signer-wallet"
+] }
+
